asomap-backend-jazzmin/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
python-decouple==3.8
pillow==10.0.1
channels==4.0.0
redis==5.0.1
factory_boy==3.3.0
Faker==20.1.0
pytest==7.4.3
pytest-django==4.7.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-jazzmin==2.6.0
django-prose-editor[sanitize]==0.16.1
beautifulsoup4==4.12.2
